draft / sate undefined error
